+++ Application/Requests/Catalog/GetOrdersRequest.cs
@@ -9,7 +9,7 @@ namespace LaptopStore.Application.Requests.Catalog
         public string UserName { get; set; }
         public string UserPhone { get; set; }
         public string UserAddress { get; set; }
-        public int TotalPrice { get; set; }
+        public long TotalPrice { get; set; }
         public string MethodPayment { get; set; }
         public string StatusOrder { get; set; }
         public bool IsPayment { get; set; }
@@ -24,7 +24,7 @@ namespace LaptopStore.Application.Requests.Catalog
         public int ProductPrice { get; set; }
         public string ProductImage { get; set; } = string.Empty;
         public int Quantity { get; set; }
-        public int TotalPrice { get; set; }
+        public long TotalPrice { get; set; }
         public bool IsRated { get; set; }
         public decimal Rate { get; set; }
     }
